# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: ShipFast Global Shipping API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /quotes:
    get:
      operationId: listQuotes
      tags: [quotes]
      summary: List all quotes
      responses:
        "200":
          description: List of quotes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Quote"
    post:
      operationId: createQuote
      tags: [quotes]
      summary: Request a shipping quote
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuoteInput"
      responses:
        "201":
          description: Quote generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quote"
  /quotes/{id}:
    get:
      operationId: getQuote
      tags: [quotes]
      summary: Get a quote by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Quote details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quote"
  /quotes/{id}/accept:
    post:
      operationId: acceptQuote
      tags: [quotes]
      summary: Accept a quote and convert to shipment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Quote accepted, shipment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shipment"

  # ── Customers ──────────────────────────────────────────────────────────────
  /customers:
    get:
      operationId: listCustomers
      tags: [customers]
      summary: List all customers
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
      responses:
        "200":
          description: List of customers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Customer"
    post:
      operationId: createCustomer
      tags: [customers]
      summary: Create a new customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerInput"
      responses:
        "201":
          description: Customer created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
  /customers/{id}:
    get:
      operationId: getCustomer
      tags: [customers]
      summary: Get customer by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Customer details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
    patch:
      operationId: updateCustomer
      tags: [customers]
      summary: Update a customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerUpdate"
      responses:
        "200":
          description: Updated customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"

  # ── Routes ─────────────────────────────────────────────────────────────────
...
  /invoices:
    get:
      operationId: listInvoices
      tags: [invoices]
      summary: List all invoices
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [draft, sent, paid, overdue, cancelled]
        - name: customerId
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: List of invoices
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Invoice"
    post:
      operationId: createInvoice
      tags: [invoices]
      summary: Create a new invoice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvoiceInput"
      responses:
        "201":
          description: Invoice created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
  /invoices/{id}:
    get:
      operationId: getInvoice
      tags: [invoices]
      summary: Get invoice by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Invoice details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
components:
  schemas:
...
    Shipment:
      type: object
      required: [id, trackingNumber, status, serviceType, originAddress, destinationAddress, senderName, recipientName, estimatedDelivery, createdAt, updatedAt]
      properties:
        id:
          type: integer
        trackingNumber:
          type: string
        status:
          type: string
          enum: [pending, processing, in_transit, out_for_delivery, delivered, cancelled, on_hold]
        serviceType:
          type: string
          enum: [standard, express, overnight, freight, international]
        originAddress:
          type: string
        originCity:
          type: string
        originCountry:
          type: string
        destinationAddress:
          type: string
        destinationCity:
          type: string
        destinationCountry:
          type: string
        senderName:
          type: string
        senderEmail:
          type: string
        senderPhone:
          type: string
        recipientName:
          type: string
        recipientEmail:
          type: string
        recipientPhone:
          type: string
        weight:
          type: number
        dimensions:
          type: ["string", "null"]
        declaredValue:
          type: ["number", "null"]
        shippingCost:
          type: number
        currency:
          type: string
          default: USD
        notes:
          type: ["string", "null"]
        customerId:
          type: ["integer", "null"]
        estimatedDelivery:
          type: string
        actualDelivery:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
...
    Quote:
      type: object
      required: [id, status, serviceType, originCity, originCountry, destinationCity, destinationCountry, weight, estimatedCost, currency, validUntil, createdAt]
      properties:
        id:
          type: integer
        status:
          type: string
          enum: [pending, active, accepted, expired]
        serviceType:
          type: string
          enum: [standard, express, overnight, freight, international]
        originCity:
          type: string
        originCountry:
          type: string
        destinationCity:
          type: string
        destinationCountry:
          type: string
        weight:
          type: number
        dimensions:
          type: ["string", "null"]
        estimatedCost:
          type: number
        currency:
          type: string
        estimatedDays:
          type: integer
        validUntil:
          type: string
        contactName:
          type: string
        contactEmail:
          type: string
        contactPhone:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        customerId:
          type: ["integer", "null"]
        createdAt:
          type: string
    QuoteInput:
      type: object
      required: [serviceType, originCity, originCountry, destinationCity, destinationCountry, weight, contactName, contactEmail]
      properties:
        serviceType:
          type: string
          enum: [standard, express, overnight, freight, international]
        originCity:
          type: string
        originCountry:
          type: string
        destinationCity:
          type: string
        destinationCountry:
          type: string
        weight:
          type: number
        dimensions:
          type: string
        declaredValue:
          type: number
        contactName:
          type: string
        contactEmail:
          type: string
        contactPhone:
          type: string
        notes:
          type: string
        customerId:
          type: integer

    # ── Customer ──────────────────────────────────────────────────────────────
    Customer:
      type: object
      required: [id, name, email, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        company:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        country:
          type: ["string", "null"]
        tier:
          type: string
          enum: [standard, premium, enterprise]
        totalShipments:
          type: integer
        totalSpend:
          type: number
        createdAt:
          type: string
    CustomerInput:
      type: object
      required: [name, email]
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        company:
          type: string
        address:
          type: string
        city:
          type: string
        country:
          type: string
        tier:
          type: string
          enum: [standard, premium, enterprise]
    CustomerUpdate:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        company:
          type: string
        address:
          type: string
        city:
          type: string
        country:
          type: string
        tier:
          type: string
          enum: [standard, premium, enterprise]

    # ── Route ─────────────────────────────────────────────────────────────────
...
    Invoice:
      type: object
      required: [id, invoiceNumber, status, customerId, amount, currency, dueDate, createdAt]
      properties:
        id:
          type: integer
        invoiceNumber:
          type: string
        status:
          type: string
          enum: [draft, sent, paid, overdue, cancelled]
        customerId:
          type: integer
        customerName:
          type: string
        amount:
          type: number
        currency:
          type: string
        shipmentIds:
          type: array
          items:
            type: integer
        dueDate:
          type: string
        paidAt:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    InvoiceInput:
      type: object
      required: [customerId, amount, dueDate]
      properties:
        customerId:
          type: integer
        amount:
          type: number
        currency:
          type: string
        shipmentIds:
          type: array
          items:
            type: integer
        dueDate:
          type: string
        notes:
          type: string