# lib/api-zod/src/generated/types/customerUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ShipFast Global Shipping API
 * OpenAPI spec version: 0.1.0
 */
import type { CustomerUpdateTier } from './customerUpdateTier';

export interface CustomerUpdate {
  name?: string;
  email?: string;
  phone?: string;
  company?: string;
  address?: string;
  city?: string;
  country?: string;
  tier?: CustomerUpdateTier;
}